feat: 지도 컴포넌트에서 marker 좌표 기능 추가
